+++ decorator/capitalize_decorator.rb
