+++ decorator/nameable.rb
