+++ decorator/trimmer_decorator.rb
