+++ person/classroom.rb
@@ -1,4 +1,4 @@
-require './student'
+require_relative 'student'
 
 class Classroom
   attr_accessor :label
