+++ person/teacher.rb
